refactor: WeatherMultiplierProvider 생성
