+++ backend/pokerogue/src/test/java/com/pokerogue/helper/battle/service/WeatherMultiplierProviderTest.java
@@ -1,8 +1,9 @@
-package com.pokerogue.helper.battle.data;
+package com.pokerogue.helper.battle.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
 
-import com.pokerogue.helper.battle.service.BattleMultiplier;
+import com.pokerogue.helper.battle.data.Weather;
 import com.pokerogue.helper.type.data.Type;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
@@ -10,14 +11,16 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class WeatherTest {
+class WeatherMultiplierProviderTest {
+
+    private final WeatherMultiplierProvider weatherMultiplierProvider = new WeatherMultiplierProvider();
 
     @ParameterizedTest
     @MethodSource(value = "multiplierConditions")
     @DisplayName("공격 기술 타입에 따른 배틀 날씨 배수를 구한다.")
     void getAttackTypeMultiplier(Weather weather, Type attackMoveType, BattleMultiplier expectedMultiplier) {
 
-        BattleMultiplier multiplier = weather.getBattleMultiplierByAttackMoveType(attackMoveType);
+        BattleMultiplier multiplier = weatherMultiplierProvider.getByAttackMoveType(weather, attackMoveType);
 
         assertThat(multiplier).isEqualTo(expectedMultiplier);
     }
